pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FastHasherMatcher"
version = "0.0.1"
authors = [{ name = "Emelia Smith", email = "emelia@brandedcode.com" }]
description = "A rewrite of Meta's Hasher-Matcher-Actioner using FastAPI"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
  "License :: OSI Approved :: GPLv3 License",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "fastapi[standard]",
  "pydantic-settings",
  "python-dotenv",
  "python-multipart",
  "jinja2",
  "sqlalchemy",
  "psycopg2",
  "requests",
  "threatexchange>=1.2.8",
  "uvicorn",
]

[project.optional-dependencies]
all = ["mypy", "black", "pytest", "types-requests", "types-psycopg2"]

test = ["pytest"]

[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true

[tool.setuptools]
py-modules = []

[[tool.mypy.overrides]]
module = "flask_apscheduler.*,importlib_metadata.*"
ignore_missing_imports = true

[project.urls]
"Homepage" = "https://github.com/thisismissem/fast-hasher-matcher/"
"Bug Tracker" = "https://github.com/thisismissem/fast-hasher-matcher/issues"
